requirements.txt
av
decord
deepspeed
einops
matplotlib
numpy
opencv-python
pandas
Pillow
scipy
tensorboard==2.9.0
tensorboardX==1.8
timm==0.4.5
torch>=1.2
torchaudio
torchvision
triton==1.0.0
pydicom
scikit-learn
